requirements.txt
mysql-connector-python==8.3.0
PyYAML==6.0.1